Move SquadJS API pings to a plugin
