--- core/constants.js
@@ -1,3 +1,0 @@
-const SQUADJS_API_DOMAIN = 'https://squadjs.thomas-smyth.uk';
-
-export { SQUADJS_API_DOMAIN };
