+++ plugins/squadjs-api/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "squadjs-command",
+  "version": "1.0.0",
+  "type": "module",
+  "private": true,
+  "dependencies": {
+    "axios": "^1.7.3"
+  }
+}
