+++ squad-server/package.json
@@ -3,7 +3,6 @@
   "version": "1.0.0",
   "type": "module",
   "dependencies": {
-    "axios": "^0.21.1",
     "core": "1.0.0",
     "gamedig": "^2.0.20"
   },
